Trocado tipo float para o BaseType.FLOAT para double

+++ projeto-compiladores-ifsc/src/main/java/ifsc/compiladores/projeto/LLVM/definitions/types/BaseType.java
@@ -7,7 +7,7 @@ public enum BaseType implements Fragment {
     BOOLEAN("i1"),
     CHAR("i8"),
     INT("i32"),
-    FLOAT("float");
+    FLOAT("double");
     private final String irDefinition;
 
     BaseType(String irDefinition) {
